requirements.txt
torch==2.7.0
torchvision==0.22.0
numpy==1.24.3
matplotlib==3.7.1
tqdm==4.65.0
Pillow==9.5.0
scipy==1.10.1
einops==0.6.1
tensorboard==2.13.0
pyyaml==6.0